tage-sc-l-emilio: Fix id overflow 32bits
The interface for Emilio's TAGE_SC_L had an overflow
when reaching the 32 bit limit, which is common in
large simulations.

The 'id' data type has been changed for the history
of the branch.

+++ src/cpu/pred/tage_sc_l_emilio.cc
@@ -81,7 +81,7 @@ TAGE_EMILIO::squash(ThreadID tid, void * &bp_history)
 bool
 TAGE_EMILIO::predict(ThreadID tid, Addr pc, bool cond_branch, void* &b)
 {
-    int id = tage.get_new_branch_id();
+    int64_t id = tage.get_new_branch_id();
     TageEmilioBranchInfo *bi = new TageEmilioBranchInfo();
     b = (void*)(bi);
     DPRINTF(Tage, "TAGE id: %d predict: %lx bp_history:%p\n", id, pc, b);
